requirements.txt
fastapi
uvicorn[standard]
aioredis
httpx
python-dotenv